a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,19 +8,19 @@
   */
 int main(void)
 {
-    int a = 1;
-    int b = 2;
+    long long int a = 1;
+    long long int b = 2;
     int c;
-    int next = a + b;
+    long long int next = a + b;
 
     printf("%d, %d, ", a, b);
 
     for (c = 3; c <= 50; c++)
-        {
-        printf("%d, ", next);
-        a = b;
-        b = next;
-        next = a + b;
-        }
+    {
+	printf("%d, ", next);
+    a = b;
+    b = next;
+    next = a + b;
+    }
     return (0);
 }
